hide form when distro already occurred
