+++ src/pages/DistributionsPage/DistributionsPage.tsx
@@ -1,30 +1,21 @@
 import assert from 'assert';
 import { useState } from 'react';
 
-import { zodResolver } from '@hookform/resolvers/zod';
+import { BigNumber, FixedNumber } from 'ethers';
 import uniqBy from 'lodash/uniqBy';
-import { useForm, SubmitHandler, Controller } from 'react-hook-form';
 import { useQuery } from 'react-query';
 import { useParams } from 'react-router-dom';
-import { z } from 'zod';
 
-import { FormControl, MenuItem, Select } from '@material-ui/core';
-
-import { ApeTextField, LoadingModal } from 'components';
+import { LoadingModal } from 'components';
 import { useContracts } from 'hooks';
 import useConnectedAddress from 'hooks/useConnectedAddress';
-import { Box, Panel, Button, Text } from 'ui';
+import { Box, Panel, Text } from 'ui';
 import { SingleColumnLayout } from 'ui/layouts';
 
 import { AllocationsTable } from './AllocationsTable';
-import { usePreviousDistributions, getEpochData } from './queries';
-import { useSubmitDistribution } from './useSubmitDistribution';
-import type { SubmitDistribution } from './useSubmitDistribution';
-
-import type { Awaited } from 'types/shim';
-
-export type QueryResult = Awaited<ReturnType<typeof getEpochData>>;
-export type Gift = Exclude<QueryResult['token_gifts'], undefined>[0];
+import { DistributionForm } from './DistributionForm';
+import { getEpochData } from './queries';
+import type { Gift } from './queries';
 
 export function DistributionsPage() {
   const { epochId } = useParams();
@@ -43,7 +34,9 @@ export function DistributionsPage() {
   );
 
   const [form1Amount, setForm1Amount] = useState<number>(0);
-  const [vault1Id, setVault1Id] = useState<string>();
+  const [vault1Id, setVault1Id] = useState<string>(
+    String(epoch?.circle?.organization?.vaults?.[0]?.id)
+  );
 
   if (isIdle || isLoading) return <LoadingModal visible />;
 
@@ -56,6 +49,7 @@ export function DistributionsPage() {
 
   const totalGive = epoch.token_gifts?.reduce((t, g) => t + g.tokens, 0) || 0;
 
+  // reformat gift data into a structure that's easier for subcomponents to use
   const usersWithReceivedAmounts = uniqBy(
     epoch.token_gifts?.map((g: Gift) => g.recipient),
     'id'
@@ -68,9 +62,20 @@ export function DistributionsPage() {
   }));
 
   const vaults = epoch.circle?.organization.vaults || [];
-  const tokenName = vaults.find(v => v.id === Number(vault1Id))?.symbol;
+  const vault1TokenName = vaults.find(v => v.id === Number(vault1Id))?.symbol;
 
-  // TODO if distribution already happened, show summary instead of form
+  let totalAmount = form1Amount;
+  let tokenName = vault1TokenName;
+
+  // if distribution already happened, show summary instead of form
+  const dist = epoch?.distributions[0];
+  if (dist) {
+    totalAmount = FixedNumber.from(dist.total_amount.toPrecision(30))
+      .mulUnsafe(dist.pricePerShare)
+      .divUnsafe(FixedNumber.from(BigNumber.from(10).pow(dist.vault.decimals)))
+      .toUnsafeFloat();
+    tokenName = dist?.vault.symbol;
+  }
 
   return (
     <SingleColumnLayout>
@@ -81,18 +86,28 @@ export function DistributionsPage() {
         <Text variant="sectionHeader" normal>
           {epoch?.circle?.name}: Epoch {epoch?.number}
         </Text>
-        <SubmitForm
-          epoch={epoch}
-          users={usersWithReceivedAmounts}
-          setAmount={setForm1Amount}
-          setVaultId={setVault1Id}
-          vaults={vaults}
-        />
+        <Panel nested css={{ mt: '$lg' }}>
+          {dist ? (
+            <>
+              <Text>There was a distribution already</Text>
+              <Text>{dist.created_at}</Text>
+            </>
+          ) : (
+            <DistributionForm
+              epoch={epoch}
+              users={usersWithReceivedAmounts}
+              setAmount={setForm1Amount}
+              setVaultId={setVault1Id}
+              vaults={vaults}
+            />
+          )}
+        </Panel>
+
         <Box css={{ mt: '$lg' }}>
           {totalGive && epoch.token_gifts ? (
             <AllocationsTable
               users={usersWithReceivedAmounts}
-              totalAmountInVault={form1Amount}
+              totalAmountInVault={totalAmount}
               tokenName={tokenName}
               totalGive={totalGive}
             />
@@ -111,194 +126,5 @@ export function DistributionsPage() {
         </Box>
       </Panel>
     </SingleColumnLayout>
-  );
-}
-
-const DistributionFormSchema = z.object({
-  amount: z.number(),
-  selectedVaultId: z.number(),
-});
-
-type DistributionForm = z.infer<typeof DistributionFormSchema>;
-
-type SubmitFormProps = {
-  epoch: QueryResult;
-  users: (Gift['recipient'] & { received: number })[];
-  setAmount: (amount: number) => void;
-  setVaultId: (vaultId: string) => void;
-  vaults: { id: number; symbol: string }[];
-};
-
-/**
- * Displays a list of allocations and allows generation of Merkle Root for a given circle and epoch.
- * @param epochId string
- * @returns JSX.Element
- */
-export function SubmitForm({
-  epoch,
-  users,
-  setAmount,
-  setVaultId,
-  vaults,
-}: SubmitFormProps) {
-  const [loadingTrx, setLoadingTrx] = useState(false);
-
-  const submitDistribution = useSubmitDistribution();
-
-  const {
-    data: prev,
-    isLoading,
-    isError,
-  } = usePreviousDistributions(epoch?.circle?.id);
-
-  const circle = epoch?.circle;
-
-  const { handleSubmit, control } = useForm<DistributionForm>({
-    resolver: zodResolver(DistributionFormSchema),
-  });
-
-  const onSubmit: SubmitHandler<DistributionForm> = async (value: any) => {
-    assert(epoch?.id && circle);
-    setLoadingTrx(true);
-    const vault = circle.organization?.vaults?.find(
-      v => v.id === Number(value.selectedVaultId)
-    );
-    assert(vault);
-
-    const gifts = users.reduce((ret, user) => {
-      ret[user.address] = user.received;
-      return ret;
-    }, {} as Record<string, number>);
-
-    const userIdsByAddress = users.reduce((ret, user) => {
-      ret[user.address.toLowerCase()] = user.id;
-      return ret;
-    }, {} as Record<string, number>);
-
-    const submitDTO: SubmitDistribution = {
-      amount: value.amount,
-      vault,
-      gifts,
-      userIdsByAddress,
-      previousDistribution: prev?.find(d => d.vault_id === vault.id),
-      circleId: circle.id,
-      epochId: epoch.id,
-    };
-
-    try {
-      await submitDistribution(submitDTO);
-      setLoadingTrx(false);
-    } catch (e) {
-      console.error('DistributionsPage.onSubmit:', e);
-      setLoadingTrx(false);
-    }
-  };
-
-  if (isLoading) return <Box>Loading...</Box>;
-
-  const pageMessage = () => {
-    if (!epoch) return `Sorry, epoch was not found.`;
-
-    if (isError)
-      return 'Sorry, there was an error retrieving previous distribution data.';
-  };
-
-  const message = pageMessage();
-  if (message) return <Box>{message}</Box>;
-
-  return (
-    <form onSubmit={handleSubmit(onSubmit)}>
-      <Box css={{ display: 'flex', justifyContent: 'center', pt: '$lg' }}>
-        <Box css={{ mb: '$lg', mt: '$xs', mr: '$md', minWidth: '15vw' }}>
-          <FormControl fullWidth>
-            <Box
-              css={{
-                color: '$text',
-                fontSize: '$4',
-                fontWeight: '$bold',
-                lineHeight: '$shorter',
-                marginBottom: '$md',
-                textAlign: 'center',
-              }}
-            >
-              Select Vault
-            </Box>
-            <Controller
-              name="selectedVaultId"
-              control={control}
-              render={({
-                field: { onChange, value },
-                fieldState: { error },
-              }) => {
-                return (
-                  <>
-                    <Select
-                      value={value || ''}
-                      label="Vault"
-                      error={!!error}
-                      disabled={loadingTrx}
-                      onChange={({ target: { value } }) => {
-                        onChange(value);
-                        setVaultId(String(value));
-                      }}
-                    >
-                      {vaults.map(vault => (
-                        <MenuItem key={vault.id} value={vault.id}>
-                          {vault.symbol}
-                        </MenuItem>
-                      ))}
-                    </Select>
-                    {error && (
-                      <Text
-                        css={{
-                          fontSize: '$3',
-                          lineHeight: '$shorter',
-                          fontWeight: '$semibold',
-                          color: '$red',
-                          textAlign: 'center',
-                          paddingTop: '$sm',
-                        }}
-                        className="error"
-                      >
-                        {error.message}
-                      </Text>
-                    )}
-                  </>
-                );
-              }}
-            />
-          </FormControl>
-        </Box>
-        <Box>
-          <Controller
-            name={'amount'}
-            control={control}
-            render={({ field: { onChange, value }, fieldState: { error } }) => (
-              <ApeTextField
-                type="number"
-                error={!!error}
-                helperText={error ? error.message : null}
-                value={value}
-                disabled={loadingTrx}
-                onChange={({ target: { value } }) => {
-                  onChange(Number(value));
-                }}
-                onBlur={({ target: { value } }) => {
-                  setAmount(Number(value));
-                }}
-                label="Total Distribution Amount"
-              />
-            )}
-          />
-        </Box>
-      </Box>
-      <Box css={{ display: 'flex', justifyContent: 'center' }}>
-        <Button color="red" size="medium" disabled={loadingTrx}>
-          {loadingTrx
-            ? `Transaction Pending...`
-            : `Submit Distribution to Vault`}
-        </Button>
-      </Box>
-    </form>
   );
 }
